Make index of a time wavelet an int64

+++ C++/time/haar_basis.hpp
@@ -17,7 +17,7 @@ class DiscConstantScalingFn : public ScalingFn<DiscConstantScalingFn> {
   constexpr static bool continuous = false;
   constexpr static const char *name = "DCS";
 
-  explicit DiscConstantScalingFn(DiscConstantScalingFn *parent, int index,
+  explicit DiscConstantScalingFn(DiscConstantScalingFn *parent, long long index,
                                  Element1D *support)
       : ScalingFn<DiscConstantScalingFn>({parent}, index, {support}) {
     assert(!support->phi_disc_const_);
@@ -47,7 +47,7 @@ class HaarWaveletFn : public WaveletFn<HaarWaveletFn> {
  public:
   constexpr static const char *name = "Haar";
 
-  explicit HaarWaveletFn(HaarWaveletFn *parent, int index,
+  explicit HaarWaveletFn(HaarWaveletFn *parent, long long index,
                          SparseVector<DiscConstantScalingFn> &&single_scale)
       : WaveletFn({parent}, index, std::move(single_scale)) {}
